skyfocus: create `SkyfocusIntegrationTest` as a Java integration test.
focus_test takes a while to run now, and running a debugger over it isn't as straightforward as a Java integration test. It'll make testing subsequent CLs easier.

This CL starts by moving the smoke test to the Java version. The Skyfocus preparation and execution steps are also moved into `CommandEnvironment` and `BuildTool` respectively, so we don't need to change `BlazeRuntimeWrapper` which mimics `BlazeCommandDispatcher`.

PiperOrigin-RevId: 622759922
Change-Id: I404f30dd957bfea521ddb2af980531bfd641f841