Support profiling in an virtual thread.
Instead of using `Thread.currentThread().threadId()` for the `threadId` in the `Profiler`, we maintain a map of `threadId` to `laneId` for virtual threads so a same lane in the UI can be re-used for different virtual threads.

This change requires using `Profiler#profileableVirtualThreadFactory()` to create VirtualThread which can use `Profiler`.

PiperOrigin-RevId: 622814148
Change-Id: I5a6bad039c95b287456597dceb158b487f61fc86