[7.2.0] Implement StartBuild (#21871)
If `--experimental_remote_output_service` is set, Bazel issues a
StartBuild RPC to the output service.

`StartBuildResponse.initial_output_path_contents` still needs to be
handled in a follow-up change.

Highlights:
- Add a new class `BazelOutputService` which is exclusive to
`RemoteOutputService`. Only `startBuild` is implemented by this PR.
Other methods will be implemented in follow-up PRs.
- Add `--experimental_remote_output_service_output_path_prefix` flag
which can be used to set the value of
`StartBuildRequest.output_path_prefix` (defaults to empty string).
- Reuse `RemoteModule#createChannel` method to create
`ReferenceCountedChannel` for `BazelOutputService`. Since there is no
server capabilities associated with this channel, the parameter is
changed to nullable.
- In `RemoteExecutionService#downloadOutputs`, the code skip downloading
outputs if `BazelOutputService` is available, instead it calls
`stageArtifacts` (although, it currently does nothing).

Tests will be added once a MVP is done.

Working towards #21630.

Closes #21674.

PiperOrigin-RevId: 617224074
Change-Id: I872ed06387e90182de2c7c69175989e9a0f663ce

Closes #21738.