treat rule toolchain reqs as query dependencies
With this change, as far as "query" is concerned: rules with toolchain-type requirements now have an implicit dependency on these toolchain types. (In Starlark rule definitions, these are specified using the `toolchains` parameter of the `rule()` function)

RELNOTES: toolchain type requirements of rules are now treated as implicit dependencies in `bazel query`. This includes those defined via the `toolchains` parameter of starlark `rule()` definitions.
PiperOrigin-RevId: 623197554
Change-Id: I3f2f546a6a19d5bfb94407b56c4d4bf48b3c0c77