feat(cc toolchain): support external_include_paths on macos
The external_include_paths feature on mac enables specifying
include paths locally with -I and with -isystem when the project
is included as an external repo.

This makes it possible to set -Werror against your own headers
for development without propagating this to consumers of your libraries.

For some reason this was only previously enabled on linux.