Emit "blaze dump --memory" data in JSON.
Just as easy to eyeball, but a bit easier to parse.

You might wonder why GSON is only used to quote strings and not to dump the whole JSON object. It's because I wanted the output to be easy to eyeball. GSON conveniently has a JsonWriter.setFormattingStyle() to emit prettified JSON, but that's not available in the latest version of GSON on Maven Central.

RELNOTES: None.
PiperOrigin-RevId: 624962535
Change-Id: If488399a2f83c36793d1515e85f78e532bda7409