Reduce memory consumption of `TrieArtifact` in `ActionInputMap`.
This can greatly reduce memory use when actions with many tree artifact inputs are in flight. Note that the memory use scales with the number of *unexpanded* tree artifacts - in other words, the cost is present even if each tree contains only one file.

There are three optimizations:

1. Do not create a `TrieArtifact` node for the terminal tree artifact. This is now possible because nested tree artifacts are prohibited (special thanks to tjgq@). This helps both by reducing the total number of `TrieArtifact` instances and reducing the shallow heap cost of a single instance from 24 to 16 bytes (since the final `TreeArtifactValue` is stored directly in the map instead of in a field).
2. Intern path segment strings.
3. Use `CompactHashMap` instead of `HashMap`.

Each of these optimizations yields significant benefit.

This data structure was added in unknown commit as a way to efficiently look up the metadata of tree artifact children without storing the whole expansion. I've added some Javadoc to the class to justify its continued existence.

PiperOrigin-RevId: 625365148
Change-Id: Ifa7d63decfacbaab060d907dfbda4531b0f0bea6