Ensure PartialReevaluation cycle node's deps are fetched if necessary
Previously, in case when the cycle node supports partial reevaluation, all of cycle nodes deps are not fetched if `env.ensurePreviouslyRequestedDepsFetched()` is not called, which will cause further execution to fail. This is a very uncommon case, so it was never caught.

On the other hand, if cycle node does not support partial reevaluation, `env.ensurePreviouslyRequestedDepsFetched()` is just a no-op, meaning this change does not introduce side effect.

PiperOrigin-RevId: 625423771
Change-Id: Iec275e7c6a45ade18f7fd0daab44b15e8d3bbe16