Migrate `AvailableXcodesInfo`, `XcodeProperties`, and `XcodeVersionRuleData` to builtin Starlark (with Starlark-consistent names).
The names of these things being very similar is likely to be confusing in the future, but we can address that after we've migrated them to apple_support. Fortunately they're only really used to communicate between the Xcode rules internally, so refactoring them shouldn't have many downstream effects.

This change does not cover `XcodeConfigInfo`, since it's a bit more complicated.

PiperOrigin-RevId: 625652603
Change-Id: Ic3c024f4b4f60c8fa93f68aa61e939215c965227