Migrate `XcodeConfigInfo` to Starlark.
Unlike the other providers, this one is very public API, used by any rule that needs to access the user's desired minimum OS version or the Xcode version used in the build. While the provider is initialized by passing all of the min OS and SDK version numbers, it only defines _functions_ to read them back out. To preserve this API during the Starlark migration, we have to make the new provider likewise return functions that capture the values that were passed in at the time of construction.

Once it's moved to apple_support, we can explore refactoring this to make it a more "normal" provider, and supply other APIs to interpret the values inside it.

PiperOrigin-RevId: 625718058
Change-Id: Ibe40774203d193e0aa4e06b2c39b2c5767a7d7f6