Do not cache action keys in actions.
This saves a bit of analysis-time memory for all builds by removing a field from all action instances. It saves a lot of execution-time memory for builds with `--use_action_cache` (the default) because we're not retaining the key strings.

For a typical `Action` on a single build, `Action#getKey` is called:

* Case 1: there is no existing action cache entry - once when writing.
* Case 2: there is an existing action cache entry, but it is a miss - twice (once when checking, once when writing).
* Case 3: there is an action cache hit - once when checking.

There are some other call sites like `Actions#canBeShared` and the `aquery` command, but these aren't hot spots.

To avoid doubling the number of key computations for case 2, we store the key in the `ActionCacheChecker.Token` for reuse when writing the cache entry. This allows the key string to be collected as soon as `ActionExecutionFunction` is done for that action.

There is a downside for incremental builds where an action's inputs change but its configured target (and hence `Action` instance) does not. Now we won't have a cached key from the previous build, but this seems like an acceptable tradeoff and is precisely the same tradeoff we committed to when clearing `NestedSet#memo`, see b/314282963#comment6 and b/314282963#comment9.

PiperOrigin-RevId: 625873158
Change-Id: I3cdd610e9301bc2f1889ffbe575305ec68dbd490