Move generating the --script_path file contents to RunCommandLine
This is a follow up to a review comment on a previous cl, and something I think
stands well on its own for code clarity.

We are forgoing some of the shared code we had from ScriptUtil (shared with
CommandFailureUtils), however that abstraction arguably wasn't pulling its
weight and made the code harder to follow and modify. With this change things
are much more direct. This also makes the issues with the windows code more
obvious.

PiperOrigin-RevId: 626005594
Change-Id: Iddd19dba1f09369c14eebe7ee4e0dbc78d4352d9