Rework the logic for writing virtual inputs, in an attempt to fix a race condition observed at Google.
I suspect that the race condition is due to LocalSpawnRunner writing virtual inputs (specifically, the PathActionInputs created by BinTools) in a non-atomic manner, thus racing with other spawn runners writing the same input.

It's also not clear to me that, for ParamFileActionInputs, it suffices to handle the dynamic execution case by providing a per-branch suffix. This rests on the assumption that no ParamFileActionInputs are shared between actions, but while this might be true today, I'd rather not rely on it to hold indefinitely. Instead, let's always use a globally unique suffix, adapting the preexisting logic in SandboxHelpers.

PiperOrigin-RevId: 626006017
Change-Id: I2c6d81a4db046d7f85832e7675835141ed21d027