Replace `ModuleExtensionResolutionEvent` with a Skyframe graph lookup
Retaining `Postable`s across the graph increases the memory usage of every dependent Skyframe node. Instead, we keep the required information in `SkyValue`s and obtain them from the "done" nodes of the graph in `BazelLockFileModule#afterCommand`.

The `BazelModuleResolutionEvent` is kept for now, but will be removed entirely with the new lockfile format.

Work towards #20369