Allow macro and target names to overlap
Previously, we had symbolic macros and targets occupy the same namespace in a package, and reported an error whenever there was a conflict between a macro and a target. This failed to consider that macros frequently (and as per style guidance, *should*) define a "main target", i.e. a target whose name is the same as the string `name` arg passed to the macro's instantiation.

This CL splits the macro and target namespaces. The relevant test cases are also inverted.

Note that the existence of a macro named "foo" will still prevent the implicit creation of an input file target named "foo". This is because we plan on allowing macro labels to be passed as inputs to other macros for the purpose of an undeclared inputs check, and we don't want to implicitly conflate this usage with an unrelated input file (see comment in beforeBuild()).

Also update relevant test cases to use text block syntax because yay for text blocks.

Work toward #19922.

PiperOrigin-RevId: 626104981
Change-Id: Id65afaadb59dc49d88621f9a9abfd8029f12f557