Add a compile build variable to indicate memprof is used.
We would like to set a custom define during linkstamp from crosstool features if memprof_optimize is specfied AND memprof_profile_path expands to a valid artifact. The memprof_profile_path variable is not exposed to the linkstamp action today. Since we only need to detect whether the memprof profile is used, we introduce a new variable "is_using_memprof" similar to fission, pic etc instead of reusing the existing profile var.

PiperOrigin-RevId: 627488520
Change-Id: I6a36b337d42033baa201f9a7080eb27d0f953c48