Ignore convenience symlinks in glob expansion
It's slightly unfortunate that this require moving symlink-related
options to an options scope that applies to query (because query _also_
needs to know to ignore the symlinks when expanding globs).

It's also slightly unfortunate that the behaviour of the symlink-related
flags is really about _changing_ the existence of symlinks - the
existence of the "ignore" value is somewhat problematic as it means that
evaluation of globs may vary depending on the outcomes of previous
builds' options, but this better (and more niche) than the current state
which is that a first build and subsequent build with the _same_ options
may give different results (because the first build may produce files in
bazel-bin which are then picked up as inputs by the second build).