Add the concept of "generic" rules to our Starlark's doc-gen.
The documentation makes a distinction between "language-specific" and
"language-agnostic" (or "generic") rules.

It is an error for these to be different between rules within a "family"
(if one is generic, they must all be generic).

However, there is currently no way to mark a rule as "generic" in
Starlark.

So we check the "type" string that is parsed from the "doc-struct" and
if it's generic, we flag it is as such. Generic rules always have their
type set to "other" so there's no issue with overloading the string in
this way.

PiperOrigin-RevId: 628120344
Change-Id: I52e89448a653bd17b5501d42b6449c3fa8ad1d24