Enable memoization in LeafObjectCodec.
This will be used to serialize SkyKeys that need to then be looked up
in Skyframe during deserialization. To avoid the complexity of having to
add a 2nd layer of continuations to codec implementations or making them
blocking, they will use LeafObjectCodec and be immediate by construction.

Since LeafObjectCodec is always MEMOIZE_AFTER, it can never have object
cycles and can use an abbreviated form of memoization which doesn't
require memoization IDs.

PiperOrigin-RevId: 628404708
Change-Id: I8095828e3d939c5dd6743d3012e906e5c5ef31f8