Add support for remote_files for http_archive
This issue adds support necessary to tackle
https://github.com/bazelbuild/bazel-central-registry/issues/1566

Add two new attributes to http_archive: remote_file_urls and
remote_file_integrity.

The purpose of these two attributes is to allow files to effectively be
overlaid ontop of an http_archive. The goal of such functionality would
be useful for BCR since the BUILD & WORKSPACE files need no longer be
stored as patch files.

This means we could probably deprecate `build_file` since that could be
referenced as a file:// url in the remote_file_urls attribute.

Co-authored-by: Mark Williams <markrwilliams@google.com>