Replace CppLinkAction in tests with SpawnAction
This makes it possible to run the same tests over Starlark version of CppLinkAction. Starlark rules create general SpawnActions.

The change was mostly done by search and replace and is a no-op.

Implement getLinkCommandLine so that it casts command line from the spawn action. This approach work both before and after the code is in Starlark.

Remove getLinkCommandLineForTesting from CppLinkAction. It is is not needed anymore.

Implement getAction(mnemonic) in BuildViewTest case. This is a stable way to retrieve a specific action created in Starlark.

PiperOrigin-RevId: 630056081
Change-Id: I3c322d405a860bfa035da879270af42b81aa3fcc