Avoid Windows short paths in Bzlmod
When a `PathFragment` is created on Windows and one of its segments looks like a Windows short path, additional file I/O is performed to resolve it to its long form. Certain Bzlmod-managed repository names could look like such paths, which potentially slows down analysis. Instead, prepend a `v` to the version of multi-version modules to avoid this.

Requires bumping the lockfile version as canonical label literals may be embedded in lockfiles.

Closes #22219.

PiperOrigin-RevId: 630468932
Change-Id: I54d58ac23e779cda50301301f16109db27d11306