[7.2.0] Correctly handle ParamFileActionInput when writing to the compact execution log. (#22247)
A ParamFileActionInput isn't present in the InputMetadataProvider and
doesn't exist in the filesystem if the respective action is running
remotely. Therefore, the isInputDirectory call as previously written
would always crash on these inputs (which is reported as a warning, not
an error; this is arguably a bad idea and will be fixed in a followup).

Fixes #21820.

PiperOrigin-RevId: 631012628
Change-Id: Ie7eb039d51b2c1465f04a18a432623b3df8e25a4