[7.2.0] Treat an IOException while logging a spawn as an error instead of a warning. (#22253)
Although failing to write a log does not in principle affect the ability
to build successfully, treating it as a warning hides problems and makes
debugging them difficult (see e.g.
https://github.com/bazelbuild/bazel/issues/21820). Spawn log
implementations must be able to inspect the spawn inputs even if they're
not present on disk (when building without the bytes) by using in-memory
data.

PiperOrigin-RevId: 631042310
Change-Id: Iabe48e524dcd77bc434037c3396956f30174f57f