Do not delay `TargetCompleteEvents` with coverage
Baseline coverage artifacts are now requested in `CompletionFunction` to ensure that they are built before the `TargetCompleteEvent` is generated. This makes it unnecessary to delay sending these events until after a Skymeld build has had the chance to request all coverage artifacts directly, which could only be done after the analysis & execution phase, thus delaying events until the end of the build.

Fixes #21475

Closes #22238.

PiperOrigin-RevId: 631414420
Change-Id: Idc77b6f5c8b5b775e6c69e35c5563f63b3bf974f