[7.2.0] Replace most Bzlmod events with a Skyframe graph lookup (#22265)
Retaining `Postable`s across the graph increases the memory usage of
every dependent Skyframe node.

`ModuleExtensionResolutionEvent` is replaced with a lookup of the "done"
Skyframe nodes for extension evaluation in
`BazelLockFileModule#afterCommand`.

`RootModuleFileFixupEvent` is replaced with direct storage in
`SingleExtensionEvalValue`. The validation of repos imported from an
extension is moved into a separate `SkyFunction`. This simplifies `bazel
mod tidy`, which no longer needs to swallow certain exceptions, and
allows cache hits when only the (invalid) imports of an extension are
modified even without the lockfile.

The `BazelModuleResolutionEvent` is kept for now, but will become
obsolete with the new lockfile format.

Work towards #20369

Closes #22058.

PiperOrigin-RevId: 628213907
Change-Id: I8ba19f5151a8183be5051c8d9280f93476db2272

---------

Co-authored-by: Fabian Meumertzheim <fabian@meumertzhe.im>