Teach "bazel dump --memory=" to dump data about every Skyframe node.
The syntax is `bazel dump --memory=full,<summary|count|bytes>` and the output is a JSON map where the canonical value of the SkyKey in question is the key and the memory dump of the associated SkyValue is the value.

The value is either a pair of (count,bytes) or a map from class to either a count of objects or bytes. This mirrors the output of `--memory=<shallow|transitive|deep>`.

This change also tweaks the object graph traverser a bit: it now counts every reference to an interned object fully as opposed to ignoring some and counting the first occurrence of some others. This seems to be a more faithful representation of reality.

RELNOTES: None.
PiperOrigin-RevId: 631741487
Change-Id: I9e3a866a0cab94ea82cbbd3e7c6558a47e076ee4