Rename LinkerInput to LegacyLinkerInput(s) and LinkerInputs.LibraryToLink to LibraryInputs
There are also CcLinkingContext.LinkerInputs and a LibraryToLink class. The same name caused a lot of confusion and made the code really difficult to explain/document. The newly-called LegacyLinkerInputs are used internally in C++ code and don't make it to public interfaces. They will eventually be completely removed.

This is mostly automatic rename done in an IDE and is a no-op.

PiperOrigin-RevId: 631777380
Change-Id: I755a49943b4ca3f6c37104ede6b360e52c7c6617