[7.2.0] Store remote registry file hashes in the lockfile (#22296)
`MODULE.bazel`, `bazel_registry.json` and `source.json` files obtained
from remote registries are stored in the repository cache and their
hashes are collected in the lockfile. This speeds up incremental module
resolutions, such as after adding a new `bazel_dep`.

Yanked versions are not stored in the lockfile. Their handling will be
part of a follow-up PR.

Implements part of
https://docs.google.com/document/d/1TjA7-M5njkI1F38IC0pm305S9EOmxcUwaCIvaSmansg/edit
Work towards #20369

Closes #21901.

PiperOrigin-RevId: 631195852
Change-Id: I35c30af7f9c3626bdbcb04c85b8c2502eeaafd3e

Co-authored-by: Fabian Meumertzheim <fabian@meumertzhe.im>