Automated rollback of commit a40fcbe45be1224e300d005d010e501a02a542cf.
*** Reason for rollback ***

b/339498986 - breaks BEP for test commands

*** Original change description ***

Refactor event handling related to metrics collection and cleanup after commands.

The end goal is to be able to report Skyframe statistics for commands that touch Skyframe but don't do builds. To this end, MetricsCollector had to be rigged up such that it reports Skyframe statistics in that case.

Previously, that relied on Blaze emiting SkyframeGraphStatsEvent in BuildTool.stopRequest(), but "query" doesn't call that. That event is (confusingly) also emitted in BlazeRuntime.afterCommand(), but...

***

PiperOrigin-RevId: 632174724
Change-Id: Ie667d6b083158fad15180c9742310e661d1301fd