Implement prefix naming rule for symbolic macros
A symbolic macro named "foo" can declare targets that either have the same name, "foo" (a main target), or the name followed by an underscore and non-empty suffix as in "foo_bar". It cannot declare targets that have just the underscore ("foo_") or that are not suffixes of the macro name.

The same naming restrictions also apply to submacros, except that macro "foo" may not declare a main submacro named "foo". (We should probably do away with that particular exclusion in the future, to avoid problems when a main target migrates from being a rule to being a submacro.)

Note that for exports_files, this precludes macro "foo" from having a target named "foo.txt". This may be a problem when the main target has an implicit output that is the name plus an extension. We may have to relax the restriction accordingly in the future.

Package.java
- Make checkForExistingTargetName()'s new entry point checkTargetName(), and likewise for macros.
- Add precondition check for addEnvironmentGroup(), which was already guarded by a soft error in the caller.

SymbolicMacroTest.java
- Migrate macroCanDefineMainTargetOfSameName to earlier in the file and rename it to macroTargetName_canBeJustNameForMainTarget.
- Add associated test cases for other valid and invalid names, and factor out a setup helper for them.
- Test that macro namespaces are still non-exclusive with targets created outside the macro.
- Test above-mentioned exports_files behavior.
- Test that submacros respect naming rules, including the main submacro mis-behavior we aim to correct.
- Replace "$" delimiter with "_" where appropriate.

Work toward #19922.

PiperOrigin-RevId: 632560579
Change-Id: I1fabcc4f0874f4c28db521c11ab536e3cb875b07