[7.2.0] Add support for remote_files for http_archive (#22325)
This issue adds support necessary to tackle
https://github.com/bazelbuild/bazel-central-registry/issues/1566

Add two new attributes to http_archive: remote_file_urls and
remote_file_integrity.

The purpose of these two attributes is to allow files to effectively be
overlaid ontop of an http_archive. The goal of such functionality would
be useful for BCR since the BUILD & WORKSPACE files need no longer be
stored as patch files.

This means we could probably deprecate `build_file` since that could be
referenced as a file:// url in the remote_file_urls attribute.

Co-authored-by: Mark Williams <markrwilliams@google.com>

CC @fmeum

Closes #22155.

PiperOrigin-RevId: 632594203
Change-Id: I6310093482c5c58537ed6dbe4ff90bafdbd696ff

Commit
https://github.com/bazelbuild/bazel/commit/e01509f1a5bbfac96e705635cfe77ed57c1e9ac9

Co-authored-by: Farid Zakaria <fmzakari@google.com>