PathFragment: test Windows mixed separator
On Windows, if a user were to specify a PathFragment with mixed
separator such as

```
bazel build --profile='C:/foo\\bar' //...
```

then the profile file will be written as `bar` and recorded in
BuildToolLogs event as so.

Provide a test that demonstrates the current behavior to discuss whether
this is "correct" or not.