Second attempt at https://github.com/bazelbuild/bazel/commit/a40fcbe45be1224e300d005d010e501a02a542cf that reports Skyframe statistics for commands that touch Skyframe but do not build.
The trouble the aforementioned change caused was caused by a chain of three events separated by about 8 years:

1. unknown commit special-cased `bazel test` so that it ignores `BuildCompleteEvent`. This was done so that the BEP correctly reports the exit code of `bazel test`.
2. unknown commit added a new `BuildMetrics` event and made it a child event of `BuildCompleteEvent` (but crucially not of `TestingCompleteEvent`).
3. https://github.com/bazelbuild/bazel/commit/a40fcbe45be1224e300d005d010e501a02a542cf made reporting `BuildMetricsEvent` reliant on parent-child relationships because it made it be posted after `BuildCompletingEvent` instances: it's posted in response to `BuildPrecompleteEvent` / `CommandPrecompleteEvent` and now that event is posted after `BuildCompletingEvent`s.

Therefore (3) made `bazel test` not emit `BuildMetricsEvent` anymore. The fix (as superficial as it is) is to add a child-parent relationship between `BuildMetricsEvent` and `TestingCompleteEvent` and also `NoTestsFound`. It is not made a child of `NoBuildRequestFinishedEvent` so that this change remains as close to the original one as possible and because it's not a problem: invocations that don't do any building don't have a build event stream.

RELNOTES: None.
PiperOrigin-RevId: 633216223
Change-Id: I26258561f07de17c8dc77031acb31d27b6038554