Add support for remote_files to bzlmod
This is a continuation of #22155 that adds the newly added
'remote_files' attribute for http_archive to the bzlmod functionality.

The end goal is to then update BCR to this new functionality to overlay
files rather than use patch files when providing MODULE/WORKSPACE/BUILD
files.

https://github.com/bazelbuild/bazel-central-registry/issues/1566 has a
good discussion of the rationale.