Add command profile to the BEP under a fixed name
If users specify a custom profile location with `--profile`, the file is added to the BES under its basename. This makes it harder for BES consumers to pick out the profile. Instead, always announce it under its default name `command.profile.gz` or `command.profile` if uncompressed. This additionally allows consumers to detect compression based on just the file name.

Closes #22345.

PiperOrigin-RevId: 633485833
Change-Id: I07fff9032d7dda34654e7a1acf9586f58c06a35b