Move the implementation of `bazel dump --memory=` to the .buildtool package.
This is in preparation for dumping the same information at the end of regular build commands akin to `--profile`; in order to do that, it needs to be done at the end of the command so hard-wiring it into `DumpCommand` just doesn't cut it.

I considered moving it to the build.skyframe package but it would have been more complicated, since the logic does entail looking at objects that the build.skyframe package doesn't know about (aspects, configured targets, etc.) so I'd have had to add another layer of abstraction, which did not look that appealing.

The game plan is to dump the data into a file and reference it through BEP.

RELNOTES: None.
PiperOrigin-RevId: 634232183
Change-Id: I2559a2a838f9208602dcfaafd9709b9286ad306f