Rewrite CcLinkingHelper.convertLibraryToLinkListToLinkerInputList to Starlark
Return None from LibraryToLink (pic_)object_files. That semantics is needed for the reset of the linking implementation.

Expose LegacyLinkerInput fields to internal Starlark implementation. Generally users shouldn't be able to get their hands on those structures.

Expose constructors of LegacyLinkerInput to cc_internal, for C++ internal consumption only.

Expose missing fields on LibraryToLink to Starlark, but only as a guarded APIs. Some of those fields weren't exposed before and we need
to evaluate if they are truly needed, before exposing them to all the users.

PiperOrigin-RevId: 634393984
Change-Id: Id31597a3b87f48566d5734d61fbb2d9ba5050e7e