feat(cc toolchain): support external_include_paths on macos
The external_include_paths feature enables specifying include paths locally with -I and with -isystem when the project is included as an external repo.

This makes it possible to set -Werror against your own headers for development without propagating this to consumers of your libraries.

For some reason this was only previously enabled on linux. I do not see a reason for not also enabling it on macos.

Closes #21949.

PiperOrigin-RevId: 634394452
Change-Id: I0df91b2a2c9b4a5bd52b22e64f99ea0745cb9759