Properly handle interrupts observed during Starlark options parsing.
Previously we were handling such an interrupt the same as an actual failure with the option (e.g. a package loading error in the package defining the option), which resulted in an exit code of `2` and a failure detail of `STARLARK_OPTIONS_PARSE_FAILURE`.

Now we correctly use an exit code of `8` and a failure detail of `INTERRUPTED`.

PiperOrigin-RevId: 634564032
Change-Id: I628c9b4abc43c5795af967436869b4a070001348